requirements.txt
openai~=0.27.2
requests~=2.28.2
python-dotenv~=1.0.0
py-cord~=2.4.0